+++ templates/javascript/controllers/login.js
@@ -7,20 +7,20 @@
  * Manages authentication to any active providers.
  */
 angular.module('<%= scriptAppName %>')
-  .controller('LoginCtrl', function ($scope, simpleLogin, $location) {
+  .controller('LoginCtrl', function ($scope, Auth, $location<% if( hasPasswordProvider ) { %>, $q, Ref, $timeout<% } %>) {
     <% if( hasOauthProviders ) { %>$scope.oauthLogin = function(provider) {
       $scope.err = null;
-      simpleLogin.login(provider, {rememberMe: true}).then(redirect, showError);
+      Auth.$authWithOAuthPopup(provider, {rememberMe: true}).then(redirect, showError);
     };
 
     $scope.anonymousLogin = function() {
       $scope.err = null;
-      simpleLogin.anonymousLogin({rememberMe: true}).then(redirect, showError);
+      Auth.$authAnonymously({rememberMe: true}).then(redirect, showError);
     };
 
     <% } %><% if( hasPasswordProvider ) { %>$scope.passwordLogin = function(email, pass) {
       $scope.err = null;
-      simpleLogin.passwordLogin({email: email, password: pass}, {rememberMe: true}).then(
+      Auth.$authWithPassword({email: email, password: pass}, {rememberMe: true}).then(
         redirect, showError
       );
     };
@@ -34,11 +34,43 @@ angular.module('<%= scriptAppName %>')
         $scope.err = 'Passwords do not match';
       }
       else {
-        simpleLogin.createAccount(email, pass, {rememberMe: true})
+        Auth.$createUser({email: email, password: pass})
+          .then(function () {
+            // authenticate so we have permission to write to Firebase
+            return Auth.$authWithPassword({email: email, password: pass}, {rememberMe: true});
+          })
+          .then(createProfile)
           .then(redirect, showError);
       }
+
+      function createProfile(user) {
+        var ref = Ref.child('users', user.uid), def = $q.defer();
+        ref.set({email: email, name: firstPartOfEmail(email)}, function(err) {
+          $timeout(function() {
+            if( err ) {
+              def.reject(err);
+            }
+            else {
+              def.resolve(ref);
+            }
+          });
+        });
+        return def.promise;
+      }
     };
-    <% } %>
+
+    function firstPartOfEmail(email) {
+      return ucfirst(email.substr(0, email.indexOf('@'))||'');
+    }
+
+    function ucfirst (str) {
+      // inspired by: http://kevin.vanzonneveld.net
+      str += '';
+      var f = str.charAt(0).toUpperCase();
+      return f + str.substr(1);
+    }
+
+  <% } %>
 
     function redirect() {
       $location.path('/account');