+++ templates/javascript/routes.without.login.js
@@ -9,18 +9,11 @@
  */
 angular.module('<%= scriptAppName %>')
 
-  // configure views; the authRequired parameter is used for specifying pages
-  // which should only be available while logged in
   .config(['$routeProvider', function($routeProvider) {
     $routeProvider
       .when('/', {
         templateUrl: 'views/main.html',
         controller: 'MainCtrl'
-      })
-
-      .when('/chat', {
-        templateUrl: 'views/chat.html',
-        controller: 'ChatCtrl'
       })
 
       .otherwise({redirectTo: '/'});